Add temporary Firebase Realtime Database read functionality
This commit introduces adds functionality in TestReceive to read data from the Firebase Realtime Database, specifically retrieving and displaying item quantities from a "requests" node. The changes in the XML files add layout for each activity, and the kotlin files handle the logout functionality and the data reading from the firebase database.

Database changes are also reflected in the web app